modify#6feat(frontEnd): component 단위로 분리
