+++ frontEnd/index.html
@@ -5,11 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Oil Pocket</title>
-
     <link rel="stylesheet" href="./public/styles/index.css">
-    <script defer type="module" src="./public/index.js"></script>
+    <script defer type="module" src="./public/js/router.js"></script>
 </head>
 <body>
-    
 </body>
 </html>
