+++ frontEnd/public/js/login/view.js
