+++ frontEnd/public/js/main/view.js
