+++ frontEnd/public/js/myPage/userDetail/view.js
@@ -1,0 +1,16 @@
+import { navBarView } from "../../navbar/view.js";
+import { getUserDetailTemplate } from "./template.js";
+
+const userDetailView = () => {
+    const $userDetailContainer = document.createElement("section");
+    const $userDetailContent = document.createElement("section");
+    $userDetailContent.classList.add("main");
+    $userDetailContent.innerHTML = getUserDetailTemplate();
+
+    $userDetailContainer.appendChild(navBarView());
+    $userDetailContainer.appendChild($userDetailContent);
+
+    return $userDetailContainer;
+}
+
+export { userDetailView }
