+++ frontEnd/public/js/navbar/template.js
