+++ frontEnd/public/js/notFound/view.js
@@ -1,0 +1,5 @@
+const notFoundView = () => {
+    return ``;
+}
+
+export { notFoundView }
