+++ frontEnd/public/js/register/template.js
