+++ frontEnd/public/js/router.js
@@ -1,14 +1,12 @@
 import { mainView } from "./main/view.js";
 import { loginView } from "./login/view.js";
 import { registerView } from "./register/view.js";
-
-import { showNotFoundView } from "./views/notFoundView.js";
-import { userDetailView } from "./views/mypage/userDetailView.js";
-import { inputOilInfoView } from "./views/mypage/inputOilInfo.js";
-import { chartView } from "./views/mypage/chart.js";
-import { comparisonView } from "./views/mypage/comparison.js";
-import { historyView } from "./views/mypage/history.js";
-
+import { userDetailView } from "./myPage/userDetail/view.js";
+import { chartView } from "./myPage/chart/view.js";
+import { comparisonView } from "./myPage/comparison/view.js";
+import { inputOilInfoView } from "./myPage/inputOilInfo/view.js";
+import { historyView } from "./myPage/history/view.js";
+import { notFoundView } from "./notFound/view.js";
 
 const $body = document.querySelector("body");
 
@@ -22,7 +20,7 @@ const router = async () => {
         { path: "/chart", view: chartView },
         { path: "/comparison", view: comparisonView },
         { path: "/history", view: historyView },
-        { path: "/404", view: showNotFoundView }
+        { path: "/404", view: notFoundView }
     ];
 
     let match = routes.map(route => {
