--- frontEnd/public/views/notFoundView.js
@@ -1,3 +1,0 @@
-const showNotFoundView = `Not Found`;
-
-export { showNotFoundView }
